# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST-API service for aggregating data on user subscriptions",
        "title": "User subscription",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cost": {
            "post": {
                "description": "The total cost of all user subscriptions by service name and period is returned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Total cost of subscriptions for the period",
                "parameters": [
                    {
                        "description": "Data for requesting the total cost",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TotalPriceReqDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful receipt of the total cost",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Data validation error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Database entry not found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            }
        },
        "/subsc": {
            "get": {
                "description": "Returns a full list of subscription records. If there are no subscriptions, returns a 404 response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Getting all subscriptions",
                "responses": {
                    "200": {
                        "description": "Successful response with subscription list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.ListSubscriptionsRow"
                            }
                        }
                    },
                    "404": {
                        "description": "No subscriptions found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new subscription record in the database with validation of dates and prices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Creating a new subscription",
                "parameters": [
                    {
                        "description": "Data for creating a subscription",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateSubReqDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscription successfully created",
                        "schema": {
                            "$ref": "#/definitions/db.CreateSubscriptionRow"
                        }
                    },
                    "400": {
                        "description": "Data validation error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            }
        },
        "/subsc/{id}": {
            "put": {
                "description": "Updates the data of an existing subscription by its ID with full parameter validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Updating a subscription record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record id (ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New subscription data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateSubReqDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful updated",
                        "schema": {
                            "$ref": "#/definitions/db.UpdateSubscriptionRow"
                        }
                    },
                    "400": {
                        "description": "Data validation error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Database entry not found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a subscription record from the database by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Deleting a record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID number of the entry in the database (id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful removal",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Database entry not found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            }
        },
        "/subsc/{user_id}": {
            "get": {
                "description": "Returning information about all user subscriptions by user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Getting a subscription record by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.GetSubscriptionsRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect user_id",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Database entry not found",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.CreateSubscriptionRow": {
            "type": "object",
            "properties": {
                "end_date": {},
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "db.ListSubscriptionsRow": {
            "type": "object",
            "properties": {
                "end_date": {},
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "db.UpdateSubscriptionRow": {
            "type": "object",
            "properties": {
                "end_date": {},
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "main.CreateSubReqDTO": {
            "type": "object",
            "required": [
                "price",
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 500
                },
                "service_name": {
                    "type": "string",
                    "example": "kion"
                },
                "start_date": {
                    "type": "string",
                    "example": "05-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "9c7ae5f1-f950-4790-8201-2b45d7853bd7"
                }
            }
        },
        "main.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "request was completed successfully"
                }
            }
        },
        "main.TotalPriceReqDTO": {
            "type": "object",
            "required": [
                "end_date",
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "10-2025"
                },
                "service_name": {
                    "type": "string",
                    "example": "ivi"
                },
                "start_date": {
                    "type": "string",
                    "example": "02-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "81997f52-03eb-42ac-89d8-e55d26b09003"
                }
            }
        },
...
    }
}